+++ apps/portal-web/src/i18n/zh_cn.ts
@@ -233,6 +233,7 @@ export default {
         saveFileFail: "文件保存失败: {}",
         saveFileSuccess: "文件保存成功",
         fileSizeExceeded: "文件过大（最大{}），请下载后编辑",
+        fileFetchAbortPrompt: "获取文件 {} 操作被终止",
       },
       createFileModal: {
         createErrorMessage: "同名文件或者目录已经存在！",
@@ -413,6 +414,31 @@ export default {
       },
     },
     dashboard:{
+      overviewTable:{
+        title:"平台概览",
+        clusterName:"集群",
+        partitionName:"分区",
+        nodeCount:"节点总数",
+        runningNodeCount:"运行中节点数",
+        cpuCoreCount:"CPU核心数",
+        gpuCoreCount:"GPU卡数",
+        usageRatePercentage:"节点使用率",
+        partitionStatus:"分区状态",
+        available:"可用",
+        notAvailable:"不可用",
+      },
+      infoPanes:{
+        nodeInfo:"节点信息",
+        node:"节点",
+        resourceInfo:"资源信息",
+        core:"核",
+        running:"运行中",
+        idle:"空闲",
+        notAvailable:"不可用",
+        card:"卡",
+        job:"作业",
+        pending:"排队中",
+      },
       addEntryModal:{
         addQuickEntry:"添加快捷方式",
         cancel:"取消",
@@ -543,7 +569,7 @@ export default {
       },
     },
     _app: {
-      textExceedsLength:"终端登录欢迎提示信息过多，请减少不必要的信息输出！",
+      textExceedsLength:"终端登录提示信息过多，请减少'~/.bashrc'等文件中不必要的信息输出!",
       sshError:"无法以用户身份连接到登录节点。请确认您的家目录的权限为700、750或者755",
       sftpError:"SFTP操作失败，请确认您是否有操作的权限",
       otherError:"服务器出错啦！",