+++ apps/portal-web/src/utils/config.ts
@@ -39,10 +39,6 @@ export interface ServerRuntimeConfig {
 
   SERVER_URL: string;
 
-  HOME_TEXTS: {[hostname: string]: string };
-
-  HOME_TITLES: {[hostname: string]: string };
-
   SUBMIT_JOB_WORKING_DIR: string;
 
   SCOW_API_AUTH_TOKEN?: string;
@@ -50,8 +46,6 @@ export interface ServerRuntimeConfig {
   AUDIT_CONFIG: AuditConfigSchema | undefined;
 
   SERVER_I18N_CONFIG_TEXTS: {
-    defaultHomeTitle: I18nStringType,
-    defaultHomeText: I18nStringType,
     submitJopPromptText?: I18nStringType,
   }