+++ apps/mis-server/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @scow/mis-server
 
+## 1.3.0
+
+### Patch Changes
+
+- Updated dependencies [ec06733f9f]
+  - @scow/scheduler-adapter-protos@1.2.0
+  - @scow/config@1.3.0
+  - @scow/lib-scheduler-adapter@1.1.3
+  - @scow/lib-hook@1.0.5
+  - @scow/lib-server@1.1.2
+  - @scow/protos@1.0.5
+
+## 1.2.3
+
+### Patch Changes
+
+- 1a1189ad48: 管理系统 AllUserTable 恢复计数接口并且新增筛选参数
+- Updated dependencies [cad49a87d8]
+  - @scow/config@1.2.1
+  - @scow/lib-hook@1.0.4
+  - @scow/lib-server@1.1.1
+  - @scow/protos@1.0.4
+  - @scow/lib-scheduler-adapter@1.1.2
+
 ## 1.2.2
 
 ### Patch Changes