+++ libs/protos/scow/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @scow/protos
 
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies [ec06733f9f]
+  - @scow/grpc-api@1.4.0
+
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies [1a1189ad48]
+  - @scow/grpc-api@1.3.1
+
 ## 1.0.3
 
 ### Patch Changes