+++ libs/server/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @scow/lib-server
 
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies [ec06733f9f]
+  - @scow/config@1.3.0
+
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies [cad49a87d8]
+  - @scow/config@1.2.1
+
 ## 1.1.0
 
 ### Minor Changes